moved blob and mime type conformance to TypeInformationDefaultConstructor in Apodini target
